# vindro-django/src/vindrobackend/openapi.yaml
openapi: 3.0.3
info:
  title: Vindro Backend API
  version: "1.0"
  description: |
    REST API for Vindro Games.
    Authentication uses session cookies (set automatically after Google OAuth login).
servers:
  - url: http://localhost:8000
    description: Local dev
  - url: https://backend.vindrogames.com
    description: Production
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: sessionid
  schemas:
...
    Tournament:
      type: object
      properties:
        id:               { type: string, format: uuid }
        name:             { type: string }
        slug:             { type: string }
        tournament_type:  { type: string, example: world_cup }
        status:           { type: string, enum: [upcoming, active, completed] }
        description:      { type: string }
        image_url:        { type: string }
        start_date:       { type: string, format: date-time }
        end_date:         { type: string, format: date-time }
        groups_end_date:  { type: string, format: date-time }
        bracket_start_date: { type: string, format: date-time }
    TournamentWithFormat:
      allOf:
        - $ref: '#/components/schemas/Tournament'
        - type: object
          properties:
            format:
              type: object
              nullable: true
              properties:
                groups:  { type: object, description: "Static group/team structure" }
                bracket: { type: object, description: "Static bracket structure" }
    TournamentResults:
      type: object
      properties:
        groups_results:   { type: object }
        bracket_results:  { type: object }
        last_updated:     { type: string, format: date-time, nullable: true }
...
    TournamentPool:
      type: object
      properties:
        id:                   { type: string, format: uuid }
        tournament_id:        { type: string, format: uuid }
        name:                 { type: string }
        description:          { type: string }
        is_public:            { type: boolean }
        current_member_count: { type: integer }
        created_by:           { type: string, nullable: true, description: "Username of creator" }
    LeaderboardEntry:
      type: object
      properties:
        play_id:   { type: string, format: uuid }
        play_name: { type: string }
        user:
          type: object
          properties:
            id:       { type: integer }
            username: { type: string }
            avatar:   { type: string, nullable: true }
        score:  { type: integer }
        status: { type: string }
...
paths:
...
  /api/tournament/:
    get:
      tags: [Tournament]
      summary: List all tournaments
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [upcoming, active, completed] }
          description: Filter by status
      responses:
        "200":
          description: List of tournaments
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Tournament' }
  /api/tournament/{tournament_id}/:
    get:
      tags: [Tournament]
      summary: Tournament detail with format (groups + bracket structure)
      parameters:
        - in: path
          name: tournament_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Tournament with format
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/TournamentWithFormat' }
        "404":
          description: Not found
  /api/tournament/{tournament_id}/results/:
    get:
      tags: [Tournament]
      summary: Live results (groups standings + bracket outcomes)
      parameters:
        - in: path
          name: tournament_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Current results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/TournamentResults' }

  # ---------------------------------------------------------------------------
  # Tournament Plays
  # ---------------------------------------------------------------------------
...
  /api/tournament/plays/{play_id}/submit/:
    post:
      tags: [Tournament Plays]
      summary: Submit (lock) a play
      security:
        - sessionAuth: []
      parameters:
        - in: path
          name: play_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Play submitted
        "400":
          description: Already submitted

  # ---------------------------------------------------------------------------
  # Tournament Pools
  # ---------------------------------------------------------------------------
...
  /api/tournament/pools/{pool_id}/leaderboard/:
    get:
      tags: [Tournament Pools]
      summary: Pool leaderboard (plays ranked by score)
      security:
        - sessionAuth: []
      parameters:
        - in: path
          name: pool_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      pool: { $ref: '#/components/schemas/TournamentPool' }
                      leaderboard:
                        type: array
                        items: { $ref: '#/components/schemas/LeaderboardEntry' }